Crate par_core

par-core

A wrapper for various parallelization library for Rust. This crate currently supports

chili

rayon

Disable parallelization.

Usage

If you are developing a library, you should not force the parallelization library, and let the users choose the parallelization library.

Final application

If you are developing a final application, you can use cargo feature to select the parallelization library.

chili

[dependencies]
par-core = { version = "1.0.1", features = ["chili"] }

rayon

[dependencies]
par-core = { version = "1.0.1", features = ["rayon"] }

Disable parallelization

[dependencies]
par-core = { version = "1.0.1", default-features = false }

Library developers

If you are developing a library, you can simply depend on par-core without any features. Note: To prevent a small mistake of end-user making the appplication slower, par-core emits a error message using a default feature. So if you are a library developer, you should specify default-features = false.

[dependencies]
par-core = { version = "1.0.1", default-features = false }

Functions

join